refactor: 1. remove BasePass; 2. move passes under the emulation folder
