+++ src/braket/aws/aws_device.py
@@ -49,7 +49,7 @@ from braket.emulation.emulator import Emulator
 from braket.emulation.local_emulator import LocalEmulator
 from braket.parametric.free_parameter import FreeParameter
 from braket.parametric.free_parameter_expression import _is_float
-from braket.passes.circuit_passes import AnkaaRxValidator, AriaMSValidator
+from braket.emulation.passes.circuit_passes import AnkaaRxValidator, AriaMSValidator
 from braket.pulse import ArbitraryWaveform, Frame, Port, PulseSequence
 from braket.pulse.waveforms import _parse_waveform_from_calibration_schema
