+++ src/braket/emulation/passes/__init__.py
@@ -11,5 +11,4 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from braket.passes.base_pass import BasePass, ProgramType  # noqa: F401
-from braket.passes.validation_pass import ValidationPass  # noqa: F401
+from braket.emulation.passes.validation_pass import ProgramType, ValidationPass  # noqa: F401
